feat: support pod management policy for statefulsets
This commit extends the spec of the `ThanosCompact`, `ThanosReceive`,
`ThanosRuler` and `ThanosStore` CRDs with a new `podManagementPolicy`
field which gets propagated to the generated statefulsets.

`podManagementPolicy` defaults to 'OrderedReady` which has the
"drawback" of requiring a manual intervention when the statefulset
roll-out gets stuck due to a misconfiguration (for instance wrong
tolerations or image location) even if the user fixes the spec later on.
When set to `Parallel`, the Kubernetes statefulset controller will
delete the "bad" pod when the spec gets updated.

Note that the workaround will work until the `MaxUnavailableStatefulSet`
feature goes GA.

https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#pod-management-policies
https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#forced-rollback
https://github.com/kubernetes/kubernetes/issues/67250

Signed-off-by: Simon Pasquier <spasquie@redhat.com>