*: Generate all YAML config from Go (#390)
* *: Generate all YAML config from Go

This commit does a few things,

- Adds a `goconfig` module to the root of the repo which contains Go
methods that can generate all of the kubebuilder scaffolding YAML
configuration defined in config/ (except the actual CRDs themselves).
This can also be imported downstream later on for users maintaining
their configs fully in Go. Also ensures sample configuration will not
"rot"
- Separate magefiles/ dir into its own go module, and adds a mage
config:generate target to generate the config/ directories and YAML
except CRDs
- Modify `manifests` target in makefile to not only call controller-gen
to generate the CRDs, but also mage config:generate to generate the full
config dir. This creates a very streamlined process for how config is
maintained in this operator.

Effectively the root of all config is `goconfig`+`controller-gen`. `make
manifests` generates the YAML in config, which is `kustomizable`. We
then use the `make build-installer` target which runs kustomize on that
generated YAML and outputs `bundle.yaml` for single line easy YAML
installs. And then we have `make build-chart` as the final layer, which
uses the bundle.yaml, and using kubebuilder helm plugin, generates a
helm chart.

This flow is a bit complex, but actually allows for a lot of flexible
install scenarios + keeps config in-line with Go code.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

* Update prom-op in e2e

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

* Move to config dir

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

* Clean up controller labels

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>

---------

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>